CLI Updates - made relative paths more robust
