--- source/README.md
@@ -1,1 +1,0 @@
-All of your project code will go here
